Тип данных метода trimmedMean() изменен на long

+++ src/SalesManager.java
@@ -17,8 +17,8 @@ public class SalesManager {
         return max;
     }
 
-    public int trimmedMean() {
-        int trimmedSum = 0;
+    public long trimmedMean() {
+        long trimmedSum = 0;
         Arrays.sort(sales);
         for (int i = 1; i < (sales.length - 1); ++i){
             trimmedSum += sales[i];
